Add a few comments the typescript to make it more understandable.
Also move saving the sidebar scroll position into another file.
